Add API usage instructions to Readme.md file Move the mealDetail.js file to the containers folder since it is not a presentational Component
